+++ swift_browser_ui_frontend/src/entries/index_app_factory.js
@@ -2,11 +2,15 @@ import { createApp } from "vue";
 
 import { i18n } from "@/common/i18n";
 
+import checkIDB from "@/common/idb_support";
+
 import { applyPolyfills, defineCustomElements } from "csc-ui/dist/loader";
 import { vControl } from "@/common/csc-ui-vue-directive";
 
 import CFooter from "@/components/CFooter.vue";
-import LanguageSelector from "@/components/CLanguageSelector.vue";
+import MainToolbar from "@/components/MainToolbar.vue";
+
+import "@/css/prod.scss";
 
 applyPolyfills().then(() => {
   defineCustomElements();
@@ -17,11 +21,14 @@ export function newApp(name, data, Component) {
     name: name,
     components: {
       CFooter,
-      LanguageSelector,
+      MainToolbar,
     },
     data: data,
     created() {
       document.title = this.$t("message.program_name");
+    },
+    mounted: function() {
+      checkIDB().then(result => this.idb = result);
     },
     methods: {
       setCookieLang: function() {